% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_rf.R
\name{build_rf}
\alias{build_rf}
\title{Build RandomForest model}
\usage{
build_rf(data, classifier)
}
\arguments{
\item{data}{A data frame containing element geochemistries of known origin}

\item{classifier}{The name/number of the column which contains the data to train the model (e.g. eruption, volcano)}
}
\value{
A tephrochron object containing the dataset, the randomForest model, and a range of evaluation objects
}
\description{
Build a RandomForest model
}
\examples{
model.rf <- (data, classifier);


}
