% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rf.R
\name{run_rf}
\alias{run_rf}
\title{Use a RandomForest model to classify tephra}
\usage{
run_rf(model.rf, new.data)
}
\arguments{
\item{data}{A data frame containing element geochemistries of unknown origin}
}
\value{
An additional column in the data frame stating the most likely origin for the tephra, based on the model
}
\description{
Use a RandomForest model to classify tephra
}
\examples{
data.classified <- (model.rf, new.data);


}
