% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_model.R
\name{select_model}
\alias{select_model}
\title{Build a numerical model to classify tephra using major element geochemistry}
\usage{
select_model(data, classifier)
}
\arguments{
\item{data}{A data frame containing element geochemistries of unknown origin}

\item{classifier}{The name/number of the column which contains the data to train the model (e.g. eruption, volcano)}
}
\value{
Confusion matrices for a variety of models,
}
\description{
Build a numerical model to classify tephra using major element geochemistry
}
\examples{
data.classified <- (data, model.rf);


Runs multiple numerical models, and allows you to choose the most effective
}
