% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splom.R
\name{splom}
\alias{splom}
\title{Scatterplot matrix}
\usage{
splom(data, colx, coly, colour, symbol)
}
\arguments{
\item{data}{A data frame containing element geochemistries}

\item{colx}{The first column containing geochemical data}

\item{coly}{The final column containing geochemical data}

\item{colour}{The variable which sets the plotting colour (e.g. sample id)}

\item{symbol}{The variable which sets the plotting symbol (e.g. site, volcano)}
}
\value{
A scatterplot matrix showing selected element geochemistries
}
\description{
Create a scatterplot matrix showing full element geochemistry
}
\examples{
plot <- biplot(data, 4, 10, Sample_ID, Site);
}
