% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{splom_select}
\alias{splom_select}
\title{Scatterplot matrix of selected elements}
\usage{
splom_select(data, cols, colour, symbol)
}
\arguments{
\item{data}{A data frame containing element geochemistries}

\item{cols}{The names of columns containing selected element geochemistries}

\item{colour}{The variable which sets the plotting colour (e.g. sample id)}

\item{symbol}{The variable which sets the plotting symbol (e.g. site, volcano)}
}
\value{
A scatterplot matrix showing selected element geochemistries
}
\description{
Create a scatterplot matrix showing selected element geochemistries
}
\examples{
plot <- splom_select(data, cols = c("SiO2","K2O","Na2O"), Sample_ID);

data$K2O_Na2O <- data$K2O / data$Na2O
plot <- splom_select(data, cols = c("SiO2","K2O_Na2O"), Sample_ID);
}
