% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_plot.R
\name{strat_plot}
\alias{strat_plot}
\title{Stratigraphic plot}
\usage{
strat_plot(data, ymin, ymax, concs, colour)
}
\arguments{
\item{data}{A data frame containing glass shard concentration data}

\item{ymin}{The lower depth of the sample}

\item{ymax}{The upper depth of the sample}

\item{concs}{Glass shard concentrations}

\item{colour}{The variable which sets the plotting colour (e.g. ash fall)}
}
\value{
A plot of glass shard concentration against depth
}
\description{
Create a plot showing tephra glass shard concentration against depth
}
\examples{
plot <- strat_plot(data, Lower_m, Upper_m, Conc_sg, Fall);
}
