% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAS_plot.R
\name{TAS}
\alias{TAS}
\title{TAS plot}
\usage{
TAS()
}
\value{
An empty TAS diagram
}
\description{
Call a blank TAS plot (I have no idea why you'd want to do this but to each their own), using fields of different magma composition after Le Maitre (2002). Based on code by John Stevenson: https://all-geo.org/volcan01010/2015/01/easily-plot-magma-compositions-tas-diagrams-in-python/
}
\details{
Le Maitre RW (2002) Igneous rocks: IUGS classification and glossary of terms: recommendations of the International Union of Geological Sciences Subcommission on the Systematics of igneous rocks, 2nd ed. Cambridge University Press, Cambridge
}
\examples{
TAS()
}
